+++ examples/java-matter-controller/java/src/com/matter/controller/commands/pairing/PairAddressPaseCommand.kt
@@ -30,9 +30,9 @@ class PairAddressPaseCommand(controller: ChipDeviceController, credsIssue: Crede
     PairingModeType.ADDRESS_PASE_ONLY,
     PairingNetworkType.NONE
   ) {
-  override fun onPairingComplete(errorCode: Long) {
+  override fun onPairingComplete(errorCode: Int) {
     logger.log(Level.INFO, "onPairingComplete with error code: $errorCode")
-    if (errorCode == 0L) {
+    if (errorCode == 0) {
       setSuccess()
     } else {
       setFailure("onPairingComplete failure")
