+++ src/controller/java/src/chip/devicecontroller/DeviceAttestationDelegate.java
@@ -31,6 +31,5 @@ public interface DeviceAttestationDelegate {
    * @param attestationInfo Attestation information for the device, null is errorCode is 0.
    * @param errorCode Error code on attestation failure. 0 if succeed.
    */
-  void onDeviceAttestationCompleted(
-      long devicePtr, AttestationInfo attestationInfo, long errorCode);
+  void onDeviceAttestationCompleted(long devicePtr, AttestationInfo attestationInfo, int errorCode);
 }
